refactor(gfn): split games.ts into focused modules behind facade
Extract cache helpers, app mapping, catalog browse/panels, and library
fetch into separate modules while preserving the public games.ts API for
existing IPC and index call sites.
